+++ src/Illuminate/Foundation/Http/Middleware/VerifyCsrfToken.php
@@ -9,6 +9,7 @@ use Illuminate\Contracts\Foundation\Application;
 use Illuminate\Contracts\Support\Responsable;
 use Illuminate\Cookie\CookieValuePrefix;
 use Illuminate\Cookie\Middleware\EncryptCookies;
+use Illuminate\Foundation\Http\Middleware\Concerns\ExcludesPaths;
 use Illuminate\Session\TokenMismatchException;
 use Illuminate\Support\Arr;
 use Illuminate\Support\InteractsWithTime;
@@ -16,7 +17,8 @@ use Symfony\Component\HttpFoundation\Cookie;
 
 class VerifyCsrfToken
 {
-    use InteractsWithTime;
+    use InteractsWithTime,
+        ExcludesPaths;
 
     /**
      * The application instance.
@@ -31,13 +33,6 @@ class VerifyCsrfToken
      * @var \Illuminate\Contracts\Encryption\Encrypter
      */
     protected $encrypter;
-
-    /**
-     * The URIs that should be excluded from CSRF verification.
-     *
-     * @var array<int, string>
-     */
-    protected $except = [];
 
     /**
      * The globally ignored URIs that should be excluded from CSRF verification.
@@ -115,24 +110,13 @@ class VerifyCsrfToken
     }
 
     /**
-     * Determine if the request has a URI that should pass through CSRF verification.
+     * Get the URIs that should be excluded.
      *
-     * @param  \Illuminate\Http\Request  $request
-     * @return bool
+     * @return array
      */
-    protected function inExceptArray($request)
+    public function getExcludedPaths()
     {
-        foreach (array_merge($this->except, static::$neverVerify) as $except) {
-            if ($except !== '/') {
-                $except = trim($except, '/');
-            }
-
-            if ($request->fullUrlIs($except) || $request->is($except)) {
-                return true;
-            }
-        }
-
-        return false;
+        return array_merge($this->except, static::$neverVerify);
     }
 
     /**
